Use enum types for (most) closed set packet fields

+++ src/CorePluginAPI/Game/Types/Skills/ESkill.cs
@@ -1,6 +1,6 @@
 ﻿namespace QuantumCore.API.Game.Types.Skills;
 
-public enum ESkill : uint
+public enum ESkill : byte
 {
     // Warrior Body
     ThreeWayCut = 1,
